refactor(archive-preview): clean imports visibility and dead helpers
